Rename the v1.management.commands.utils module
This is not a management command so it shouldn't appear in the list of commands.
